added shelf item component, and imported dispatch and use effect
